Migrate tiledb_array_schema_evolution_t struct definition into C API folder.
