+++ CMakeLists.txt
@@ -294,6 +294,7 @@ list(APPEND TILEDB_C_API_RELATIVE_HEADERS
     "${CMAKE_SOURCE_DIR}/tiledb/api/c_api/array_schema/array_schema_api_external.h"
     "${CMAKE_SOURCE_DIR}/tiledb/api/c_api/array_schema/array_type_enum.h"
     "${CMAKE_SOURCE_DIR}/tiledb/api/c_api/array_schema/layout_enum.h"
+    "${CMAKE_SOURCE_DIR}/tiledb/api/c_api/array_schema_evolution/array_schema_evolution_api_experimental.h"
     "${CMAKE_SOURCE_DIR}/tiledb/api/c_api/attribute/attribute_api_external.h"
     "${CMAKE_SOURCE_DIR}/tiledb/api/c_api/attribute/attribute_api_external_experimental.h"
     "${CMAKE_SOURCE_DIR}/tiledb/api/c_api/buffer/buffer_api_external.h"
