+++ tiledb/api/c_api/array/array_api.cc
@@ -34,6 +34,7 @@
 #include "array_api_internal.h"
 
 #include "tiledb/api/c_api/array_schema/array_schema_api_internal.h"
+#include "tiledb/api/c_api/array_schema_evolution/array_schema_evolution_api_internal.h"
 #include "tiledb/api/c_api/context/context_api_internal.h"
 #include "tiledb/api/c_api/enumeration/enumeration_api_internal.h"
 #include "tiledb/api/c_api/query/query_api_external.h"
