+++ tiledb/sm/c_api/api_argument_validator.h
@@ -56,19 +56,6 @@ inline void ensure_subarray_is_valid(const tiledb_subarray_t* p) {
 }
 }  // namespace tiledb::api
 
-inline int32_t sanity_check(
-    tiledb_ctx_t* ctx,
-    const tiledb_array_schema_evolution_t* schema_evolution) {
-  if (schema_evolution == nullptr ||
-      schema_evolution->array_schema_evolution_ == nullptr) {
-    auto st = Status_Error("Invalid TileDB array schema evolution object");
-    LOG_STATUS_NO_RETURN_VALUE(st);
-    save_error(ctx, st);
-    return TILEDB_ERR;
-  }
-  return TILEDB_OK;
-}
-
 inline int32_t sanity_check(tiledb_ctx_t* ctx, const tiledb_query_t* query) {
   if (query == nullptr || query->query_ == nullptr) {
     auto st = Status_Error("Invalid TileDB query object");
