+++ tiledb/sm/c_api/tiledb.h
@@ -57,6 +57,7 @@
  */
 #include "tiledb/api/c_api/array/array_api_external.h"
 #include "tiledb/api/c_api/array_schema/array_schema_api_external.h"
+#include "tiledb/api/c_api/array_schema_evolution/array_schema_evolution_api_experimental.h"
 #include "tiledb/api/c_api/attribute/attribute_api_external.h"
 #include "tiledb/api/c_api/buffer/buffer_api_external.h"
 #include "tiledb/api/c_api/buffer_list/buffer_list_api_external.h"
